--- .vscode/.vscode/settings.json
@@ -1,8 +1,0 @@
-{
-  "editor.tabSize": 2,
-  "editor.formatOnSave": true,
-  "editor.wordWrap": "wordWrapColumn",
-  "html.format.wrapAttributes": "force-aligned",
-  "editor.minimap.maxColumn": 80,
-  "editor.defaultFormatter": "esbenp.prettier-vscode"
-}
