+++ client/script.js
